test mailgun state against desired state
+ refactor to split state into github + mailgun state
+ use mailgun state for terraform importer
