+++ config.go
@@ -272,103 +272,6 @@ func LoadConfig(tree fs.FS) (*Config, error) {
 	}, nil
 }
 
-func (cfg *Config) DesiredGitHubState() GitHubState {
-	var state GitHubState
-
-	repoCollaborators := map[string][]GitHubRepoCollaborator{}
-
-	for _, person := range cfg.Contributors {
-		state.Members = append(state.Members, GitHubOrgMember{
-			Name:  person.Name,
-			Login: person.GitHub,
-			Role:  OrgRoleMember,
-		})
-
-		for repo, permission := range person.Repos {
-			repoCollaborators[repo] = append(repoCollaborators[repo], GitHubRepoCollaborator{
-				Login:      person.GitHub,
-				Permission: permission3to4(permission),
-			})
-		}
-	}
-
-	for _, team := range cfg.Teams {
-		ghTeam := GitHubTeam{
-			Name:        team.Name,
-			Description: sanitize(team.Purpose),
-		}
-
-		for _, member := range team.Members(cfg) {
-			ghTeam.Members = append(ghTeam.Members, GitHubTeamMember{
-				Login: member.GitHub,
-				Role:  TeamRoleMember,
-			})
-		}
-
-		for _, repo := range team.Repos {
-			ghTeam.Repos = append(ghTeam.Repos, GitHubTeamRepoAccess{
-				Name:       repo,
-				Permission: team.RepoPermission(),
-			})
-		}
-
-		state.Teams = append(state.Teams, ghTeam)
-	}
-
-	for _, repo := range cfg.Repos {
-		ghRepo := GitHubRepo{
-			Name:                repo.Name,
-			Description:         sanitize(repo.Description),
-			IsPrivate:           repo.Private,
-			Topics:              repo.Topics,
-			HomepageURL:         repo.HomepageURL,
-			HasIssues:           repo.HasIssues,
-			HasProjects:         repo.HasProjects,
-			HasWiki:             repo.HasWiki,
-			DirectCollaborators: repoCollaborators[repo.Name],
-		}
-
-		for _, protection := range repo.BranchProtection {
-			ghRepo.BranchProtectionRules = append(ghRepo.BranchProtectionRules, GitHubRepoBranchProtectionRule{
-				Pattern: protection.Pattern,
-
-				AllowsDeletions: protection.AllowsDeletions,
-
-				RequiresStatusChecks: len(protection.RequiredChecks) > 0,
-
-				// ensure there's an empty slice so comparing in tests doesn't fail
-				// with nil != []
-				RequiredStatusCheckContexts: append([]string{}, protection.RequiredChecks...),
-
-				// having no checks configured seems to force this setting to be
-				// enabled
-				RequiresStrictStatusChecks: protection.StrictChecks || len(protection.RequiredChecks) == 0,
-
-				RequiresApprovingReviews:     protection.RequiredReviews > 0,
-				RequiredApprovingReviewCount: protection.RequiredReviews,
-				DismissesStaleReviews:        protection.DismissStaleReviews,
-				RequiresCodeOwnerReviews:     protection.RequireCodeOwnerReviews,
-
-				// hardcoded defaults in github.tf
-				IsAdminEnforced:   false,
-				AllowsForcePushes: false,
-			})
-		}
-
-		for _, deployKey := range repo.DeployKeys {
-			ghRepo.DeployKeys = append(ghRepo.DeployKeys, GitHubDeployKey{
-				Title:    deployKey.Title,
-				Key:      deployKey.PublicKey,
-				ReadOnly: !deployKey.Writable,
-			})
-		}
-
-		state.Repos = append(state.Repos, ghRepo)
-	}
-
-	return state
-}
-
 func (config Config) SyncMissing(dest string) error {
 	for name, person := range config.Contributors {
 		filePath := filepath.Join(dest, "contributors", name+".yml")
